refactor: изменено размещение вспомогательных директорий
