+++ src/components/blocks/genres-list/genres-list.jsx
@@ -4,7 +4,7 @@ import classnames from "classnames";
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import * as actions from "../../../store/actions";
-import {getGenresItems} from "../../utils/utils";
+import {getGenresItems} from "../../../utils/utils";
 import MoviesList from "../movies-list/movies-list";
 
 const GenresList = ({movies, currentGenre, resetFilter, changeGenre}) => {
