+++ src/components/blocks/movie-card-tabs/movie-card-tabs.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import MovieCardNav from "./movie-card-nav/movie-card-nav";
-import {TabsItems} from "../../constants/tabs";
-import MovieCardTabContent from "../../hocs/movie-card-tab-content";
+import {TabsItems} from "../../../constants/tabs";
+import MovieCardTabContent from "../../../hocs/movie-card-tab-content";
 
 const MovieCardTabs = ({activeKey, defaultActiveKey = 1, ...currentMovie}) => {
