+++ src/components/blocks/user-block/user-block.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import {connect} from "react-redux";
-import {AuthorizationStatus} from "../../constants/auth";
+import {AuthorizationStatus} from "../../../constants/auth";
 import Tooltip from "../tooltip/tooltip";
-import {RoutePath} from "../../constants/routes";
+import {RoutePath} from "../../../constants/routes";
 import {useHistory} from "react-router-dom";
 import PropTypes from "prop-types";
 import {logout} from "../../../store/api-actions";
