+++ src/components/pages/add-review/add-review.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import MainLayout from "../../layouts/main-layout/main-layout";
 import MovieCard from "../../sections/movie-card/movie-card";
-import useMovie from "../../hooks/use-movie";
+import useMovie from "../../../hooks/use-movie";
 import {connect} from "react-redux";
 import MoviesList from "../../blocks/movies-list/movies-list";
