+++ src/components/pages/player/player.jsx
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import MainLayout from "../../layouts/main-layout/main-layout";
 import VideoPlayer from "../../sections/video-player/video-player";
 import {useHistory} from "react-router-dom";
-import useMovie from "../../hooks/use-movie";
+import useMovie from "../../../hooks/use-movie";
 import {connect} from "react-redux";
 import MoviesList from "../../blocks/movies-list/movies-list";
