+++ src/components/sections/page-header/page-header.jsx
@@ -3,7 +3,7 @@ import Logo from "../../blocks/logo/logo";
 import classnames from "classnames";
 import PropTypes from "prop-types";
 import UserBlock from "../../blocks/user-block/user-block";
-import {RoutePath} from "../../constants/routes";
+import {RoutePath} from "../../../constants/routes";
 import Link from "../../blocks/link/link";
 import {useParams} from "react-router";
