+++ src/hocs/movie-card-tab-content.jsx
